feat(hs): further modularize, introduce util module
Further modularize the Brainfuck library by creating a module for the
data type (and instances) definition, a module for the operators, and a
module for utility functions.

The modules containing the data type and the operators are exposed
through the "main" Brainfuck module, while Util is exposed in the
package to be imported directly.